requirements.txt
nonebot2>=2.3.0
nonebot-adapter-onebot>=2.4.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
matplotlib>=3.8.0
Pillow>=10.0.0
apscheduler>=3.10.0
nonebot2[fastapi]
